fix: update source-handling for new build process
Note: this gates all configuration-related scripts behind file-
detection. I.e., we attempt to detect the script to-be-run before
executing it. This enables support for both the new and old
configuration scripts to function (allowing the TC to seamlessly
use either the legacy script or the updated scripts as-sensible).

Also note: this is horrible to maintain; the issue is that tooling
like the TC assumes the locations of things in the repository are
static and unchanging (even though we know the locations and naming
of many things are suboptimal); as a result moving anything breaks
downstream tooling constantly.

A longer-term fix would be to provide a stable interface (e.g., a
wrapper `Makefile` with targets we keep up-to-date; or a manifest
file that simply lists the locations of things), so tooling can
target the stable interface instead of hard-coding paths.

cf. opencbdc-tx#249

Signed-off-by: Sam Stuewe <stuewe@mit.edu>